cleanup: merged assign_roles & assign_identifiers management commands
